% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divsa.R
\name{divsa}
\alias{divsa}
\title{TCGA Data Import, Visualisation and Statistical Analysis}
\usage{
divsa(table, phenotype, cancer_type, meta_lev1, meta_lev2, meta_col, patient)
}
\arguments{
\item{table}{Full path to the input file}

\item{phenotype}{The specific phenotype for comparison, given as the name of the column from table}

\item{cancer_type}{The type of the tumor from the 'type' column}

\item{meta_lev1}{First level of data from chosen metadata column}

\item{meta_lev2}{Second level of data from chosen metadata column}

\item{meta_col}{The chosen metadata column for additional filtering of cases}

\item{patient}{Ids of patients to analyze from the 'bcr_patient_barcode' column}
}
\value{
A boxplot on phenotype distributions and statistical comparison of medians using Wilcox test
}
\description{
Package imports data from TCGA-CDR spreadsheet of Supplementary Data File 1 from Liu et al. 2018
paper. It allows to phenotype data to compare around specific patients or patients with specific
tumor type exhibiting one other additional feature. The phenotype, patiennts, tumor type, and
additional feature can be freely selected on the basis of the spreadsheet TCGA-CDR data.
The comparison involves boxplot visualisation on phenotype distributions as well as
statistical comparison of medians using Wilcox non-parametric test.
}
